Hasegawa Wakatini Project Documentation

Welcome to the documentation page.

Quick-Glimpse Tutorial

This tutorial just offered a quick glimpse on Julia's built-in documentation system, make sure to read the docs for more.

HasagawaWakatini — Module

Includes Domain, Operators, Parameters and Timestepper

HasagawaWakatini.Helperfunctions.ifftPlot — Function

ifftPlot(args...; kwargs...) Plot the real part of the inverse Fourier transform (IFFT) of each argument that is a complex array. This function is designed to handle multiple input arrays and plot them using the plot function from a plotting library such as Plots.jl. Non-complex arrays are plotted as-is.

Arguments

args...: A variable number of arguments. Each argument can be an array. If the array is of a complex type, its IFFT is computed, and only the real part is plotted. If the array is not complex, it is plotted directly.

kwargs...: Keyword arguments that are passed directly to the plot function to customize the plot.

Usage

using FFTW, Plots

Create some sample data

x = rand(ComplexF64, 100)
y = rand(100)

Plot the real part of the IFFT of x and y directly

ifftPlot(x, y, title="IFFT Plot Example", legend=:topright)

HasagawaWakatini.Helperfunctions.Domain — Type

Box domain, that calculates spatial resolution under construction.

Contains

Lengths: $Lx$, $Ly$ (Float64)
Number of grid point: $Nx$, $Ny$ (Int64)
Spatial resolution: $dx$, $dy$ (Float64)
Spatial points: $x$, $y$ (LinRange)

$dxᵢ = 2Lₓ÷(Nₓ-1)$

Square Domain can be constructed using:
$Domain(N,L)$

Rectangular Domain can be constructed using:
$Domain(Nx,Ny,Lx,Ly)$

HasagawaWakatini.Operators.Laplacian! — Method

Non-allocating function

dn - change in field
n - the current field
K - pre allocated K = -(kx^2 + ky^2)

HasagawaWakatini.Operators.Laplacian — Method

Laplacian Operator

n - current field a - frequencies (parameter) t - time (not used)

HasagawaWakatini.Operators.∇ — Method

Short hand notation for laplacian Operator

n - current field a - frequencies (parameter) t - time (not used)